+++ src/GlobalHelper.php
@@ -175,10 +175,9 @@ class GlobalHelper
     public static function setupDefaultCache(): void
     {
         if (!isset(self::$cache)) {
-            $home = PlatformUtil::getUserDirectory();
-            $directory = $home ? $home . '/.cache' : sys_get_temp_dir();
+            $directory = PlatformUtil::getCacheDirectory();
 
-            self::setCache(new FilesystemAdapter(directory: $directory . '/castor'));
+            self::setCache(new FilesystemAdapter(directory: $directory));
         }
     }